Moved utils to their own module
Moved these filesystem utility functions out as I'll likely use
them in the main script.
